requirements.txt
ann-visualizer==2.5
base58==2.1.1
bcrypt==3.1.7
bitarray==1.2.1
fitipy==0.1.2
grpcio==1.60.0
HeapDict==1.0.1
intervaltree==3.0.2
joblib==0.14.1
more-itertools==8.2.0
numpy==1.21.4
oauthlib==3.2.2
Pillow==8.4.0
pkginfo==1.5.0.1
prometheus-client==0.7.1
pycosat==0.6.3
pynacl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
PyWavelets==1.1.1
QtPy==1.9.0
resampy==0.4.2
sphinx==2.4.0
streamlit==1.23.1
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
torch==1.13.1
torchgeometry==0.1.2
typing-extensions==4.7.1
tzdata==2024.1
tzlocal==4.3.1
ujson==1.35
urllib3==1.25.8
validators==0.20.0
wcwidth==0.2.6
zict==1.0.0
